create connection manager
