+++ src/commands/commandactivate.cpp
@@ -38,48 +38,53 @@ int CommandActivate::run(QStringList& tokens) {
     }
 
     QEventLoop loop;
-    QObject::connect(vpn.controller(), &Controller::stateChanged, &vpn, [&] {
-      if (vpn.controller()->state() == Controller::StateOff ||
-          vpn.controller()->state() == Controller::StateOn) {
-        loop.exit();
-      }
-    });
+    QObject::connect(
+        vpn.connectionManager(), &ConnectionManager::stateChanged, &vpn, [&] {
+          if (vpn.connectionManager()->state() == ConnectionManager::StateOff ||
+              vpn.connectionManager()->state() == ConnectionManager::StateOn) {
+            loop.exit();
+          }
+        });
 
-    vpn.controller()->initialize();
+    vpn.connectionManager()->initialize();
     loop.exec();
-    vpn.controller()->disconnect();
+    vpn.connectionManager()->disconnect();
 
     // If we are connecting right now, we want to wait untile the operation is
     // completed.
-    if (vpn.controller()->state() != Controller::StateOff &&
-        vpn.controller()->state() != Controller::StateOn) {
-      QObject::connect(vpn.controller(), &Controller::stateChanged, &vpn, [&] {
-        if (vpn.controller()->state() == Controller::StateOff ||
-            vpn.controller()->state() == Controller::StateOn) {
-          loop.exit();
-        }
-      });
+    if (vpn.connectionManager()->state() != ConnectionManager::StateOff &&
+        vpn.connectionManager()->state() != ConnectionManager::StateOn) {
+      QObject::connect(vpn.connectionManager(),
+                       &ConnectionManager::stateChanged, &vpn, [&] {
+                         if (vpn.connectionManager()->state() ==
+                                 ConnectionManager::StateOff ||
+                             vpn.connectionManager()->state() ==
+                                 ConnectionManager::StateOn) {
+                           loop.exit();
+                         }
+                       });
       loop.exec();
-      vpn.controller()->disconnect();
+      vpn.connectionManager()->disconnect();
     }
 
-    if (vpn.controller()->state() != Controller::StateOff) {
+    if (vpn.connectionManager()->state() != ConnectionManager::StateOff) {
       QTextStream stream(stdout);
       stream << "The VPN tunnel is already active" << Qt::endl;
       return 0;
     }
 
-    QObject::connect(vpn.controller(), &Controller::stateChanged, &vpn, [&] {
-      if (vpn.controller()->state() == Controller::StateOff ||
-          vpn.controller()->state() == Controller::StateOn) {
-        loop.exit();
-      }
-    });
-    vpn.controller()->activate(*vpn.serverData());
+    QObject::connect(
+        vpn.connectionManager(), &ConnectionManager::stateChanged, &vpn, [&] {
+          if (vpn.connectionManager()->state() == ConnectionManager::StateOff ||
+              vpn.connectionManager()->state() == ConnectionManager::StateOn) {
+            loop.exit();
+          }
+        });
+    vpn.connectionManager()->activate(*vpn.serverData());
     loop.exec();
-    vpn.controller()->disconnect();
+    vpn.connectionManager()->disconnect();
 
-    if (vpn.controller()->state() == Controller::StateOn) {
+    if (vpn.connectionManager()->state() == ConnectionManager::StateOn) {
       QTextStream stream(stdout);
       stream << "The VPN tunnel is now active" << Qt::endl;
       return 0;
