+++ src/commands/commandui.cpp
@@ -17,6 +17,7 @@
 #include "commandlineparser.h"
 #include "connectionbenchmark/connectionbenchmark.h"
 #include "connectionhealth.h"
+#include "connectionmanager.h"
 #include "constants.h"
 #include "controller.h"
 #include "feature.h"
@@ -271,8 +272,9 @@ int CommandUI::run(QStringList& tokens) {
 #ifndef Q_OS_WIN
     // Signal handling for a proper shutdown.
     SignalHandler sh;
-    QObject::connect(&sh, &SignalHandler::quitRequested,
-                     []() { MozillaVPN::instance()->controller()->quit(); });
+    QObject::connect(&sh, &SignalHandler::quitRequested, []() {
+      MozillaVPN::instance()->connectionManager()->quit();
+    });
 #endif
 
     // Font loader
@@ -307,7 +309,7 @@ int CommandUI::run(QStringList& tokens) {
         "Mozilla.VPN", 1, 0, "VPNCaptivePortal",
         MozillaVPN::instance()->captivePortalDetection());
     qmlRegisterSingletonInstance("Mozilla.VPN", 1, 0, "VPNController",
-                                 MozillaVPN::instance()->controller());
+                                 MozillaVPN::instance()->connectionManager());
     qmlRegisterSingletonInstance("Mozilla.VPN", 1, 0, "VPNUser",
                                  MozillaVPN::instance()->user());
     qmlRegisterSingletonInstance("Mozilla.VPN", 1, 0, "VPNDeviceModel",
@@ -385,8 +387,8 @@ int CommandUI::run(QStringList& tokens) {
         []() { MozillaVPN::instance()->deactivate(true); },
         Qt::DirectConnection);
 
-    QObject::connect(vpn.controller(), &Controller::readyToQuit, &vpn,
-                     &App::quit, Qt::QueuedConnection);
+    QObject::connect(vpn.connectionManager(), &ConnectionManager::readyToQuit,
+                     &vpn, &App::quit, Qt::QueuedConnection);
 
     // Here is the main QML file.
     const QUrl url(QStringLiteral("qrc:/ui/main.qml"));
@@ -399,7 +401,7 @@ int CommandUI::run(QStringList& tokens) {
     NotificationHandler* notificationHandler =
         NotificationHandler::create(&engineHolder);
 
-    QObject::connect(vpn.controller(), &Controller::stateChanged,
+    QObject::connect(vpn.connectionManager(), &ConnectionManager::stateChanged,
                      notificationHandler,
                      &NotificationHandler::showNotification);
 
@@ -410,8 +412,8 @@ int CommandUI::run(QStringList& tokens) {
     QObject::connect(&vpn, &MozillaVPN::stateChanged, &menuBar,
                      &MacOSMenuBar::controllerStateChanged);
 
-    QObject::connect(vpn.controller(), &Controller::stateChanged, &menuBar,
-                     &MacOSMenuBar::controllerStateChanged);
+    QObject::connect(vpn.connectionManager(), &ConnectionManager::stateChanged,
+                     &menuBar, &MacOSMenuBar::controllerStateChanged);
 
 #endif
 
@@ -443,8 +445,8 @@ int CommandUI::run(QStringList& tokens) {
 
 #ifdef MVPN_WEBEXTENSION
     ServerHandler serverHandler;
-    QObject::connect(vpn.controller(), &Controller::readyToQuit, &serverHandler,
-                     &ServerHandler::close);
+    QObject::connect(vpn.connectionManager(), &ConnectionManager::readyToQuit,
+                     &serverHandler, &ServerHandler::close);
 #endif
 
     KeyRegenerator keyRegenerator;
