+++ src/controllerimpl.h
@@ -8,7 +8,7 @@
 #include <QObject>
 #include <functional>
 
-#include "controller.h"
+#include "connectionmanager.h"
 
 class Keys;
 class Device;
@@ -40,11 +40,11 @@ class ControllerImpl : public QObject {
   // state terminates when the "connected" (or the "disconnected") signal is
   // received.
   virtual void activate(const InterfaceConfig& config,
-                        Controller::Reason Reason) = 0;
+                        ConnectionManager::Reason Reason) = 0;
 
   // This method terminates the VPN tunnel. The VPN client is in
   // "disconnecting" state until the "disconnected" signal is received.
-  virtual void deactivate(Controller::Reason reason) = 0;
+  virtual void deactivate(ConnectionManager::Reason reason) = 0;
 
   // This method is used to retrieve the VPN tunnel status (mainly the number
   // of bytes sent and received). It's called always when the VPN tunnel is
