+++ src/keyregenerator.cpp
@@ -4,6 +4,7 @@
 
 #include "keyregenerator.h"
 
+#include "connectionmanager.h"
 #include "constants.h"
 #include "controller.h"
 #include "feature.h"
@@ -27,7 +28,7 @@ KeyRegenerator::KeyRegenerator() {
   MozillaVPN* vpn = MozillaVPN::instance();
 
   connect(vpn, &MozillaVPN::stateChanged, this, &KeyRegenerator::stateChanged);
-  connect(vpn->controller(), &Controller::stateChanged, this,
+  connect(vpn->connectionManager(), &ConnectionManager::stateChanged, this,
           &KeyRegenerator::stateChanged);
   connect(&m_timer, &QTimer::timeout, this, &KeyRegenerator::stateChanged);
   connect(Feature::get(Feature::Feature_keyRegeneration),
@@ -51,7 +52,7 @@ void KeyRegenerator::stateChanged() {
   MozillaVPN* vpn = MozillaVPN::instance();
 
   if (vpn->state() != App::StateMain ||
-      vpn->controller()->state() != Controller::StateOff) {
+      vpn->connectionManager()->state() != ConnectionManager::StateOff) {
     logger.debug() << "Wrong state";
     return;
   }
