+++ src/localsocketcontroller.h
@@ -24,9 +24,9 @@ class LocalSocketController final : public ControllerImpl {
   void initialize(const Device* device, const Keys* keys) override;
 
   void activate(const InterfaceConfig& config,
-                Controller::Reason Reason) override;
+                ConnectionManager::Reason Reason) override;
 
-  void deactivate(Controller::Reason reason) override;
+  void deactivate(ConnectionManager::Reason reason) override;
 
   void checkStatus() override;
