+++ src/mozillavpn.cpp
@@ -8,7 +8,9 @@
 #include "addons/manager/addonmanager.h"
 #include "authenticationinapp/authenticationinapp.h"
 #include "captiveportal/captiveportaldetection.h"
+#include "connectionmanager.h"
 #include "constants.h"
+#include "controller.h"
 #include "dnshelper.h"
 #include "externalophandler.h"
 #include "feature.h"
@@ -133,7 +135,7 @@ MozillaVPN::MozillaVPN() : App(nullptr), m_private(new MozillaVPNPrivate()) {
     // If we are activating the app, let's initialize the controller and the
     // periodic tasks.
     if (state() == StateMain) {
-      m_private->m_controller.initialize();
+      m_private->m_connectionManager.initialize();
       startSchedulingPeriodicOperations();
     } else {
       stopSchedulingPeriodicOperations();
@@ -153,25 +155,26 @@ MozillaVPN::MozillaVPN() : App(nullptr), m_private(new MozillaVPNPrivate()) {
              : Navigator::NoFlags));
   });
 
-  connect(&m_private->m_controller, &Controller::readyToUpdate, this,
-          [this]() { setState(StateUpdateRequired); });
+  connect(&m_private->m_connectionManager, &ConnectionManager::readyToUpdate,
+          this, [this]() { setState(StateUpdateRequired); });
 
-  connect(&m_private->m_controller, &Controller::readyToBackendFailure, this,
-          [this]() {
+  connect(&m_private->m_connectionManager,
+          &ConnectionManager::readyToBackendFailure, this, [this]() {
             TaskScheduler::deleteTasks();
             setState(StateBackendFailure);
           });
 
-  connect(&m_private->m_controller, &Controller::readyToServerUnavailable, this,
+  connect(&m_private->m_connectionManager,
+          &ConnectionManager::readyToServerUnavailable, this,
           [](bool pingReceived) {
             NotificationHandler::instance()->serverUnavailableNotification(
                 pingReceived);
           });
 
-  connect(&m_private->m_controller, &Controller::stateChanged, this,
-          &MozillaVPN::controllerStateChanged);
+  connect(&m_private->m_connectionManager, &ConnectionManager::stateChanged,
+          this, &MozillaVPN::controllerStateChanged);
 
-  connect(&m_private->m_controller, &Controller::stateChanged,
+  connect(&m_private->m_connectionManager, &ConnectionManager::stateChanged,
           &m_private->m_statusIcon, &StatusIcon::refreshNeeded);
 
   connect(this, &MozillaVPN::stateChanged, &m_private->m_statusIcon,
@@ -180,11 +183,11 @@ MozillaVPN::MozillaVPN() : App(nullptr), m_private(new MozillaVPNPrivate()) {
   connect(&m_private->m_connectionHealth, &ConnectionHealth::stabilityChanged,
           &m_private->m_statusIcon, &StatusIcon::refreshNeeded);
 
-  connect(&m_private->m_controller, &Controller::stateChanged,
+  connect(&m_private->m_connectionManager, &ConnectionManager::stateChanged,
           &m_private->m_connectionHealth,
           &ConnectionHealth::connectionStateChanged);
 
-  connect(&m_private->m_controller, &Controller::stateChanged,
+  connect(&m_private->m_connectionManager, &ConnectionManager::stateChanged,
           &m_private->m_captivePortalDetection,
           &CaptivePortalDetection::stateChanged);
 
@@ -891,7 +894,7 @@ void MozillaVPN::mainWindowLoaded() {
 #endif
 #ifdef SENTRY_ENABLED
   SentryAdapter::instance()->init();
-  QObject::connect(controller(), &Controller::readyToQuit,
+  QObject::connect(connectionManager(), &ConnectionManager::readyToQuit,
                    SentryAdapter::instance(), &SentryAdapter::onBeforeShutdown);
 #endif
 }
@@ -999,8 +1002,9 @@ void MozillaVPN::silentSwitch() {
   // here, the connection does not work and we don't want to wait for timeouts
   // to run the silenct-switch.
   TaskScheduler::deleteTasks();
-  TaskScheduler::scheduleTask(new TaskControllerAction(
-      TaskControllerAction::eSilentSwitch, Controller::eServerCoolDownNeeded));
+  TaskScheduler::scheduleTask(
+      new TaskControllerAction(TaskControllerAction::eSilentSwitch,
+                               ConnectionManager::eServerCoolDownNeeded));
 }
 
 void MozillaVPN::refreshDevices() {
@@ -1126,8 +1130,9 @@ void MozillaVPN::update() {
   // The windows installer will stop the client and daemon before installation
   // so it's not necessary to disable the VPN to perform an upgrade.
 #ifndef MZ_WINDOWS
-  if (m_private->m_controller.state() != Controller::StateOff &&
-      m_private->m_controller.state() != Controller::StateInitializing) {
+  if (m_private->m_connectionManager.state() != ConnectionManager::StateOff &&
+      m_private->m_connectionManager.state() !=
+          ConnectionManager::StateInitializing) {
     deactivate();
     return;
   }
@@ -1153,7 +1158,8 @@ void MozillaVPN::controllerStateChanged() {
     }
   }
 
-  if (m_updating && m_private->m_controller.state() == Controller::StateOff) {
+  if (m_updating &&
+      m_private->m_connectionManager.state() == ConnectionManager::StateOff) {
     update();
   }
 
@@ -1169,7 +1175,7 @@ void MozillaVPN::heartbeatCompleted(bool success) {
   logger.debug() << "Server-side check done:" << success;
 
   if (!success) {
-    m_private->m_controller.backendFailure();
+    m_private->m_connectionManager.backendFailure();
     return;
   }
 
@@ -1285,8 +1291,9 @@ void MozillaVPN::scheduleRefreshDataTasks() {
   // server selection is implemented upon activation. See JIRA issue
   // https://mozilla-hub.atlassian.net/browse/VPN-3726 for more information.
   if (!m_private->m_location.initialized()) {
-    Controller::State st = m_private->m_controller.state();
-    if (st == Controller::StateOff || st == Controller::StateInitializing) {
+    ConnectionManager::State st = m_private->m_connectionManager.state();
+    if (st == ConnectionManager::StateOff ||
+        st == ConnectionManager::StateInitializing) {
       TaskScheduler::scheduleTask(
           new TaskGetLocation(ErrorHandler::PropagateError));
     }
@@ -1406,15 +1413,18 @@ void MozillaVPN::registerErrorHandlers() {
   ErrorHandler::registerCustomErrorHandler(
       ErrorHandler::DependentConnectionError, true, []() {
         MozillaVPN* vpn = MozillaVPN::instance();
-        if (vpn->controller()->state() == Controller::State::StateOn ||
-            vpn->controller()->state() == Controller::State::StateConfirming) {
+        if (vpn->connectionManager()->state() ==
+                ConnectionManager::State::StateOn ||
+            vpn->connectionManager()->state() ==
+                ConnectionManager::State::StateConfirming) {
           // connection likely isn't stable yet
           logger.error()
               << "Ignore network error probably caused by enabled VPN";
           return ErrorHandler::NoAlert;
         }
 
-        if (vpn->controller()->state() == Controller::State::StateOff) {
+        if (vpn->connectionManager()->state() ==
+            ConnectionManager::State::StateOff) {
           // We are off, so this means a request failed, not the
           // VPN. Change it to No Connection
           return ErrorHandler::NoConnectionAlert;
@@ -1926,7 +1936,7 @@ void MozillaVPN::registerInspectorCommands() {
         MozillaVPN::instance()
             ->captivePortalDetection()
             ->captivePortalDetected();
-        MozillaVPN::instance()->controller()->captivePortalPresent();
+        MozillaVPN::instance()->connectionManager()->captivePortalPresent();
         return QJsonObject();
       });
 
@@ -2110,7 +2120,7 @@ void MozillaVPN::registerInspectorCommands() {
   InspectorHandler::registerCommand(
       "quit", "Quit the app", 0,
       [](InspectorHandler*, const QList<QByteArray>&) {
-        MozillaVPN::instance()->controller()->quit();
+        MozillaVPN::instance()->connectionManager()->quit();
         return QJsonObject();
       });
 
@@ -2159,7 +2169,7 @@ void MozillaVPN::registerAddonApis() {
     QJSEngine* engine = QmlEngineHolder::instance()->engine();
 
     {
-      QObject* obj = MozillaVPN::instance()->controller();
+      QObject* obj = MozillaVPN::instance()->connectionManager();
       QQmlEngine::setObjectOwnership(obj, QQmlEngine::CppOwnership);
 
       QJSValue value = engine->newQObject(obj);
@@ -2213,7 +2223,7 @@ void MozillaVPN::registerExternalOperations() {
   eoh->registerExternalOperation(OpNotificationClicked, []() {});
 
   eoh->registerExternalOperation(
-      OpQuit, []() { MozillaVPN::instance()->controller()->quit(); });
+      OpQuit, []() { MozillaVPN::instance()->connectionManager()->quit(); });
 }
 
 void MozillaVPN::ensureApplicationIdExists() {
