+++ src/mozillavpn_p.h
@@ -9,6 +9,7 @@
 #include "captiveportal/captiveportaldetection.h"
 #include "connectionbenchmark/connectionbenchmark.h"
 #include "connectionhealth.h"
+#include "connectionmanager.h"
 #include "controller.h"
 #include "ipaddresslookup.h"
 #include "models/devicemodel.h"
@@ -31,6 +32,7 @@ struct MozillaVPNPrivate {
   CaptivePortalDetection m_captivePortalDetection;
   ConnectionBenchmark m_connectionBenchmark;
   ConnectionHealth m_connectionHealth;
+  ConnectionManager m_connectionManager;
   Controller m_controller;
   DeviceModel m_deviceModel;
   IpAddressLookup m_ipAddressLookup;
