+++ src/platforms/ios/ioscontroller.h
@@ -19,9 +19,9 @@ class IOSController final : public ControllerImpl {
   void initialize(const Device* device, const Keys* keys) override;
 
   void activate(const InterfaceConfig& config,
-                Controller::Reason reason) override;
+                ConnectionManager::Reason reason) override;
 
-  void deactivate(Controller::Reason reason) override;
+  void deactivate(ConnectionManager::Reason reason) override;
 
   void checkStatus() override;
