+++ src/tasks/controlleraction/taskcontrolleraction.cpp
@@ -16,10 +16,10 @@ Logger logger("TaskControllerAction");
 
 TaskControllerAction::TaskControllerAction(
     TaskControllerAction::TaskAction action,
-    Controller::ServerCoolDownPolicyForSilentSwitch serverCoolDownPolicy)
+    ConnectionManager::ServerCoolDownPolicyForSilentSwitch serverCoolDownPolicy)
     : Task("TaskControllerAction"),
       m_action(action),
-      m_lastState(Controller::State::StateOff),
+      m_lastState(ConnectionManager::State::StateOff),
       // Let's take a copy of the current server-data to activate/switch to the
       // current locations even if the settings change in the meantime.
       m_serverData(*MozillaVPN::instance()->serverData()),
@@ -38,31 +38,33 @@ TaskControllerAction::~TaskControllerAction() {
 void TaskControllerAction::run() {
   logger.debug() << "TaskControllerAction run";
 
-  Controller* controller = MozillaVPN::instance()->controller();
-  Q_ASSERT(controller);
+  ConnectionManager* connectionManager =
+      MozillaVPN::instance()->connectionManager();
+  Q_ASSERT(connectionManager);
 
-  connect(controller, &Controller::stateChanged, this,
+  connect(connectionManager, &ConnectionManager::stateChanged, this,
           &TaskControllerAction::stateChanged, Qt::QueuedConnection);
 
   bool expectSignal = false;
 
-  m_lastState = controller->state();
+  m_lastState = connectionManager->state();
 
   switch (m_action) {
     case eActivate:
-      expectSignal = controller->activate(m_serverData);
+      expectSignal = connectionManager->activate(m_serverData);
       break;
 
     case eDeactivate:
-      expectSignal = controller->deactivate();
+      expectSignal = connectionManager->deactivate();
       break;
 
     case eSilentSwitch:
-      expectSignal = controller->silentSwitchServers(m_serverCoolDownPolicy);
+      expectSignal =
+          connectionManager->silentSwitchServers(m_serverCoolDownPolicy);
       break;
 
     case eSwitch:
-      expectSignal = controller->switchServers(m_serverData);
+      expectSignal = connectionManager->switchServers(m_serverData);
       break;
   }
 
@@ -83,13 +85,14 @@ void TaskControllerAction::stateChanged() {
     return;
   }
 
-  Controller* controller = MozillaVPN::instance()->controller();
-  Q_ASSERT(controller);
+  ConnectionManager* connectionManager =
+      MozillaVPN::instance()->connectionManager();
+  Q_ASSERT(connectionManager);
 
-  Controller::State state = controller->state();
+  ConnectionManager::State state = connectionManager->state();
   if (((m_action == eActivate || m_action == eSwitch) &&
-       state == Controller::StateOn) ||
-      (m_action == eDeactivate && state == Controller::StateOff)) {
+       state == ConnectionManager::StateOn) ||
+      (m_action == eDeactivate && state == ConnectionManager::StateOff)) {
     logger.debug() << "Operation completed";
     m_timer.stop();
     emit completed();
@@ -97,12 +100,14 @@ void TaskControllerAction::stateChanged() {
 }
 
 void TaskControllerAction::checkStatus() {
-  Controller* controller = MozillaVPN::instance()->controller();
-  Q_ASSERT(controller);
-  if (controller->state() == m_lastState) {
+  ConnectionManager* connectionManager =
+      MozillaVPN::instance()->connectionManager();
+  Q_ASSERT(connectionManager);
+
+  if (connectionManager->state() == m_lastState) {
     m_timer.stop();
     emit completed();
   } else {
-    m_lastState = controller->state();
+    m_lastState = connectionManager->state();
   }
 }
