+++ tests/qml/CMakeLists.txt
@@ -39,6 +39,7 @@ target_compile_definitions(qml_tests PRIVATE MZ_DUMMY)
 target_sources(qml_tests PRIVATE
     ${MZ_SOURCE_DIR}/constants.h
     ${MZ_SOURCE_DIR}/controller.h
+    ${MZ_SOURCE_DIR}/connectionmanager.h
     ${MZ_SOURCE_DIR}/dnspingsender.cpp
     ${MZ_SOURCE_DIR}/dnspingsender.h
     ${MZ_SOURCE_DIR}/models/location.cpp
