+++ tests/unit/mocmozillavpn.cpp
@@ -35,6 +35,10 @@ ConnectionHealth* MozillaVPN::connectionHealth() const { return nullptr; }
 
 Controller* MozillaVPN::controller() const { return new Controller(); }
 
+ConnectionManager* MozillaVPN::connectionManager() const {
+  return new ConnectionManager();
+}
+
 DeviceModel* MozillaVPN::deviceModel() const { return nullptr; }
 
 Keys* MozillaVPN::keys() const { return nullptr; }
